+++ client/src/components/LogoutButton/index.js
@@ -5,7 +5,7 @@ const LogoutButton = () => {
     const { logout } = useAuth0();
 
     return (
-        <button className="btn btn-primary" onClick={() => logout({ returnTo: window.location.origin })}>
+        <button className="btn btn-primary float-right" onClick={() => logout({ returnTo: window.location.origin })}>
             Log out
         </button>
     );
