--- client/src/pages/PrimeVideo.js
@@ -1,48 +1,0 @@
-import React from 'react';
-
-const PrimeVideo = () => {
-    return (
-        <div>
-            <div>
-                <style>{'body { background-image: url(https://wallpaperaccess.com/full/2312674.jpg); }'}</style>
-            </div>
-            <div class="jumbotron jumbotron-fluid">
-                <div class="container">
-                    <h1 class="display-4 text-center">Prime Video</h1>
-                </div>
-            </div>
-            <div class="jumbotron jumbotron-fluid">
-                <div class="container m-0">
-                    <p class="lead">Top Shows</p>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Show"/></button>
-                </div>
-            </div>
-            <div class="jumbotron jumbotron-fluid">
-                <div class="container m-0">
-                    <p class="lead">Top Movies</p>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                    <button type="button" class="btn mr-1 mt-1 btn-sm rounded shadow-lg"><img src="https://via.placeholder.com/250x140" alt="Movie"/></button>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default PrimeVideo;
